test(config): split test_config.py into 6 domain files
The original tests/test_config.py was 1,519 lines covering every
config dataclass in one file, which meant any config change had to
re-load the whole thing into context. Split it along the same axis as
the dataclasses themselves:

  tests/test_config_loading.py        (10 tests) - file parse, defaults
  tests/test_config_notifications.py  (9 tests)  - legacy Discord, avatar
  tests/test_config_filesystems.py    (3 tests)  - mount path parsing
  tests/test_config_vm_containers.py  (7 tests)  - compose files, runtime
  tests/test_config_remote.py         (29 tests) - servers, ordering, margin
  tests/test_config_validation.py     (25 tests) - cross-field, edge cases

83 tests total, identical to the pre-split count. Test bodies are
moved verbatim - only the file boundaries and per-file imports change.

docs/testing.md updated with the new file count (17 files now), the
new test count after the multi-ups coverage push (364), and the
per-area breakdown listing each split file.
